move random.py to utils
