+++ test/second_q/circuit/library/test_bogoliubov_transform.py
@@ -13,7 +13,7 @@
 """Test Bogoliubov transform circuits."""
 
 from test import QiskitNatureTestCase
-from test.random import random_quadratic_hamiltonian
+from qiskit_nature.utils.random import random_quadratic_hamiltonian
 
 import numpy as np
 from ddt import data, ddt, unpack
